backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
